requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
bzip2=1.0.8=h1de35cc_0
ca-certificates=2021.7.5=hecd8cb5_1
cairo=1.16.0=h8023c5d_1
certifi=2021.5.30=py37hecd8cb5_0
ffmpeg=4.0=h01ea3c9_0
fontconfig=2.13.1=ha9ee91d_0
freetype=2.10.4=ha233b18_0
gettext=0.21.0=h7535e17_0
glib=2.69.0=hdf23fa2_0
graphite2=1.3.14=h38d11af_0
harfbuzz=1.8.8=hb8d4a28_0
hdf5=1.10.2=hfa1e0ec_1
icu=58.2=h0a44026_3
intel-openmp=2021.2.0=hecd8cb5_564
jasper=2.0.14=h636a363_1
jpeg=9b=he5867d9_2
lcms2=2.12=hf1fd2bf_0
libcxx=10.0.0=1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libopencv=3.4.2=h7c891bd_1
libopus=1.3.1=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libtiff=4.2.0=h87d7836_0
libvpx=1.7.0=h378b8a2_0
libwebp-base=1.2.0=h9ed2024_0
libxml2=2.9.12=hcdb78fc_0
llvm-openmp=10.0.0=h28b9765_0
lz4-c=1.9.3=h23ab428_0
mkl=2021.2.0=hecd8cb5_269
mkl-service=2.3.0=py37h9ed2024_1
mkl_fft=1.3.0=py37h4a7008c_2
mkl_random=1.2.1=py37hb2f4e1b_2
ncurses=6.2=h0a44026_1
numpy=1.20.2=py37h4b4dc7a_0
numpy-base=1.20.2=py37he0bd621_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
openjpeg=2.3.0=hb95cd4c_1
openssl=1.1.1k=h9ed2024_0
pcre=8.45=h23ab428_0
pillow=8.3.1=py37ha4cf6ea_0
pip=21.1.3=py37hecd8cb5_0
pixman=0.40.0=haf1e3a3_0
py-opencv=3.4.2=py37h7c891bd_1
python=3.7.10=h88f2d9e_0
readline=8.1=h9ed2024_0
setuptools=52.0.0=py37hecd8cb5_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hce871da_0
tk=8.6.10=hb0a8c7a_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
